+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-# test Changelog
+# Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Added Translit] - 2025-02-16
+
+- Initial release of the Translit extension
+- Support for standard Russian phonetic transliteration rules
+- Real-time conversion as you type
+- Copy to clipboard action